# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Access-Nex
  description: >
    Self-hosted OAuth2 / OpenID Connect provider. This document covers the
    JSON surface a client application (including a separately-hosted
    frontend — see docs/FRONTEND.md) can build against: the standard
    OIDC/OAuth2 endpoints, the /api/v1 account API, and the /api/admin
    management API. The server-rendered HTML pages (/login, /portal, /admin,
    /device) are not part of this contract — they exist for zero-setup
    browser use and can be replaced entirely by a custom frontend calling
    these APIs.
  version: "1.0"
servers:
  - url: http://localhost:8080
    description: Default local install
security:
  - sessionCookie: []
paths:
...
  /authorize:
    get:
      summary: Authorization endpoint (browser navigation, not an API call)
      description: >
        Standard OAuth2/OIDC authorization request. Not JSON — this is a
        browser-navigated flow that ends in a redirect (or a login/consent/
        2FA HTML page along the way) per RFC 6749/OIDC Core. A custom
        frontend still navigates the browser here for the OAuth dance
        itself; only account management (login, 2FA, consent history) has a
        JSON equivalent under /api/v1.
      security: []
      parameters:
        - { name: response_type, in: query, required: true, schema: { type: string, enum: [code] } }
        - { name: client_id, in: query, required: true, schema: { type: string } }
        - { name: redirect_uri, in: query, schema: { type: string } }
        - { name: scope, in: query, schema: { type: string } }
        - { name: state, in: query, schema: { type: string } }
        - { name: nonce, in: query, schema: { type: string } }
        - { name: code_challenge, in: query, schema: { type: string } }
        - { name: code_challenge_method, in: query, schema: { type: string, enum: [S256, plain] } }
        - { name: response_mode, in: query, schema: { type: string, enum: [query, form_post] } }
        - { name: prompt, in: query, schema: { type: string, enum: [none, login, consent, select_account] } }
      responses:
        "302": { description: Redirect to redirect_uri with code (or error), or to the login page }
        "200": { description: Login, 2FA, or consent HTML page }
...
  /api/v1/login:
    post:
      summary: Password login (step 1)
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { username: { type: string }, password: { type: string } }
      responses:
        "200":
          description: Signed in, or a 2FA challenge
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties: { status: { type: string, enum: [signed_in] } }
                  - type: object
                    properties:
                      status: { type: string, enum: [2fa_required] }
                      token: { type: string, description: "pass to /api/v1/login/totp or /login/webauthn/*" }
                      methods: { type: array, items: { type: string, enum: [totp, webauthn] } }
        "401": { description: Invalid credentials }
        "429": { description: Rate limited }
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: oidc_sid
      description: >
        Set by /login, /api/v1/login(+totp), or /login/webauthn/finish.
        Mutating /api/v1 and /api/admin requests additionally require the
        X-Access-Nex-Api / X-Access-Nex-Admin header respectively.
    clientBasic:
      type: http
      scheme: basic
      description: OAuth2 client_id:client_secret (confidential clients); public clients omit auth entirely
    bearerToken:
      type: http
      scheme: bearer
    dpopToken:
      type: apiKey
      in: header
      name: Authorization
      description: 'For DPoP-bound tokens: "Authorization: DPoP <token>" plus a DPoP proof header'